Split email service into EmailBuilderService
